bend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
blinker==1.7.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.2
chromadb==0.4.6
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.0
convex==0.5.0
cycler==0.11.0
dataclasses-json==0.5.13
dnspython==2.4.0
fastapi==0.99.1
Flask==3.0.1
Flask-Cors==4.0.0
flask-sqlalchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.41.0
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
humanfriendly==10.0
idna==3.4
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
langchain==0.0.240
langsmith==0.0.14
loguru==0.7.0
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.7.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.4.0
packaging==23.1
pandas==2.2.0
Pillow==10.0.0
pinecone-client==2.2.2
posthog==3.0.2
protobuf==4.24.1
pulsar-client==3.2.0
pydantic==1.10.11
Pygments==2.15.1
pyparsing==3.0.9
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-sql==1.4.3
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.4
urllib3==2.0.3
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==3.0.1
yarl==1.9.2
